+++ frontend/src/types/Company.ts
