+++ frontend/src/types/Leetcode.ts
